Update guide documentation for 0.3
Update the guide to be in sync with the changes made for 0.3.

I've rewritten the tutorial chapter, starting using the offline dependency provider to implementing your own dependency provider, focussed on getting a user with no prior solver experience from getting started quickly to eventually writing a full dependency provider with all custom types.

Some parts are completely removed (such as the continuous versions) as the new traits and structs make them unnecessary. Others i've only touched up to not be not wrong, though they could use a rewrite with the input from uv and cargo. I want to unblock the 0.3 release without starting any new projects.

I'm tempted to remove the top-level chapter pages and instead move their (remaining) content into the first page of each chapter.

We should move the book into the main pubgrub repository to prevent it from getting out of sync in the future. The only non-trivial item is the cloudflare secrets.