Add prettier and improve CI (#13)
* Add prettier and improve CI

To keep a consistent style with multiple editors, and to avoid manual reflowing of text, we add prettier as formatter for markdown and yml files.

I've intentionally left everything at default settings to rely on the work defining a good standard done by prettier's team and to avoid bikeshedding over the style.

The only non-default option is prose wrapping for markdown files, which adds the equivalent of a line length limit as we have it in rust and yaml to markdown, too. I find too long lines hard to read and soft-wrap isn't consistently available (e.g. github PR diffs become hard to read), though we could remove that option, too.

Prettier is enforced by CI. On CI, we also build the book for all PRs with an updated mdbook version and check compatibility with the latest mdbook version (non-blocking).

* Adds names to jobs

* Don't run CI twice for PRs